libsensors: switch to portable typedefs to match callbacks
Using a strict basetype only coding style is not a good idea
especially when we are dealing with callbacks which use portable
typedef like int64_t. This kind of coding style would most likely
end up in overflows at various places especially when dealing with
timestamps and sensor delays which are caused when typecasting
datatypes of different size.

Switch to portable typedefs and get rid of "long".

Change-Id: I75b9cace7602345dba9095f046292e6d4db07df4